Add files via upload
This code uses the tkinter library to create a simple GUI with two entry boxes and a button. When the user enters a long URL and clicks the "Shorten URL" button, the code uses the pyshorteners library to shorten the URL and display the shortened URL in the second entry box.

Note that you will need to install the pyshorteners library by running pip install pyshorteners in your command prompt or terminal before running this code.